Unban the jsr305 jars.
All plugins that would report clashing of this jar with the
findbugs-annotations jar can now configured to ignore this conflict
which is benign and only happens because annotation classes exist in
both jars.

As the industry seems to have decided to go with the jsr305 jar
(because the annotations jar is not as widely known), basepom no
longer enforces a policy of using the annotations jar istead of the
jsr305 jar.