Adds hubspot dependency-management plugin
This adds the plugin and hooks it into the basic set of tests.

However, the default settings are that neither dependency nor plugin
management are enforced. The rationale is that for a single module
project, requiring to use management sections for every plugin and
dependency would be a significant burden (and break existing
projects). It would be desirable for multi-module projects (central
management in one place is a good thing) so there it can be activated
through the added properties (basepom.dependency-management.dependencies
and basepom.dependency-management.plugins).

The alternative would have been to enforce checking but not fail the
build when violations exist. This would change the default fail
setting, so the plugin could not use the basic check defaults and it
would clutter perfectly fine single module builds with unnecessary
warnings.