Remove forked test process timeout in surefire and failsafe plugins
The setting of forkedProcessTimeoutInSeconds=30 (seconds) seems very short.
It means a forked test JVM is killed if it runs beyond 30 seconds.
Projects with forkCount=1 and a reasonable number of unit tests are likely to breach this limit.
The plugin default (0, meaning wait forever for the process) seems a more reasonable default.